2019-10-31: v1.0.2 : Better recurring handling using dedicated PayBatch cron scripts.
