+++ modules/gateways/payhostpaybatch.php
@@ -12,21 +12,37 @@
  *
  */
 
+require_once 'payhostpaybatch/lib/constants.php';
+require_once 'payhostpaybatch/lib/payhostsoap.class.php';
+
 if ( !defined( "WHMCS" ) ) {
     die( "This file cannot be accessed directly" );
 }
 
 use WHMCS\Database\Capsule;
 
+if ( !defined( '_DB_PREFIX_' ) ) {
+    define( '_DB_PREFIX_', 'tbl' );
+}
+
 /**
- * Global defines used for the two endpoints
+ * Check for existence of payhostpaybatch table and create if not
  */
-define( "PAYHOSTAPI", 'https://secure.paygate.co.za/payhost/process.trans' );
-define( "PAYHOSTAPIWSDL", 'https://secure.paygate.co.za/payhost/process.trans/?wsdl' );
-define( "PAYBATCHAPI", 'https://secure.paygate.co.za/paybatch/1.2/process.trans' );
-define( "PAYBATCHAPIWSDL", 'https://secure.paygate.co.za/paybatch/1.2/PayBatch.wsdl' );
-define( "PAYGATETESTID", '10011072130' );
-define( "PAYGATETESTKEY", 'test' );
+if ( !function_exists( 'createPayhostpaybatchTable' ) ) {
+    function createPayhostpaybatchTable()
+    {
+        $query = "create table if not exists `" . _DB_PREFIX_ . "payhostpaybatch` (";
+        $query .= " id INT(6) UNSIGNED AUTO_INCREMENT PRIMARY KEY, ";
+        $query .= " recordtype VARCHAR(20) NOT NULL, ";
+        $query .= " recordid VARCHAR(50) NOT NULL, ";
+        $query .= " recordval VARCHAR(50) NOT NULL, ";
+        $query .= " dbid VARCHAR(10) NOT NULL DEFAULT '1')";
+
+        return full_query( $query );
+    }
+}
+
+createPayhostpaybatchTable();
 
 /**
  * Define module related meta data
@@ -132,27 +148,23 @@ function payhostpaybatch_link( $params )
     // Check if test mode or not
     $testMode = $params['testMode'];
     if ( $testMode == 'on' ) {
-        $payHostId         = PAYGATETESTID;
-        $payBatchId        = PAYGATETESTID;
-        $payHostSecretKey  = PAYGATETESTKEY;
-        $payBatchSecretKey = PAYGATETESTKEY;
+        $payHostId        = PAYGATETESTID;
+        $payHostSecretKey = PAYGATETESTKEY;
     } else {
-        $payHostId         = $params['payHostID'];
-        $payBatchId        = $params['payBatchID'];
-        $payHostSecretKey  = $params['payHostSecretKey'];
-        $payBatchSecretKey = $params['payBatchSecretKey'];
+        $payHostId        = $params['payHostID'];
+        $payHostSecretKey = $params['payHostSecretKey'];
     }
 
     $gatewayModuleName = basename( __FILE__, '.php' );
     $html              = '';
 
     // Check if recurring payments and vaulting are allowed - if not, do not enable PayBatch
-    $recurring   = $params['payhostpaybatch_recurring'];
-    $vaulting    = $params['payhostpaybatch_vaulting'];
-    $usePayBatch = false;
-    if ( $recurring == 'on' && $vaulting == 'on' ) {
-        $usePayBatch = true;
+    $recurring = $params['payhostpaybatch_recurring'];
+    $vaulting  = $params['payhostpaybatch_vaulting'];
+    if ( $vaulting == 'on' ) {
+        $vaulting = true;
     }
+    $usePayBatch = false;
 
     // System Parameters
     $companyName                             = $params['companyname'];
@@ -189,106 +201,72 @@ function payhostpaybatch_link( $params )
     $amount       = $params['amount'];
     $currencyCode = $params['currency'];
 
-    // Check for a value stored in tblclients->cardnum : we will use this to store the card vault id
-    $vaultId = Capsule::table( 'tblclients' )
-        ->where( 'id', $params['clientdetails']['userid'] )
-        ->value( 'cardnum' );
+    // Check for a value of correct format stored in tblclients->cardnum : we will use this to store the card vault id
+    $tblpayhostpaybatch = _DB_PREFIX_ . 'payhostpaybatch';
+    $vaultId            = Capsule::table( $tblpayhostpaybatch )
+        ->where( 'recordtype', 'clientdetail' )
+        ->where( 'recordid', $params['clientdetails']['userid'] )
+        ->value( 'recordval' );
+    $vaultPattern = '/^[0-9a-z]{8}-[0-9a-z]{4}-[0-9a-z]{4}-[0-9a-z]{4}-[0-9a-z]{12}$/';
+    if ( preg_match( $vaultPattern, $vaultId ) != 1 ) {
+        $vaultId = '';
+    }
 
-    if ( isset( $vaultId ) && $vaultId != '' && $usePayBatch === true ) {
-        // A vault transaction - use PayBatch with single entry
-        $payBatchSoap = new paybatchsoap( $notifyUrl );
+    // A web payment request - use PayHost WebPayment request for redirect
 
-        if ( $currencyCode !== 'ZAR' ) {
-            die( 'ZAR is required for PayBatch. Cannot proceed.' );
-        }
+    $payhostSoap = new payhostsoap();
 
-        $data   = [];
-        $item   = [];
-        $item[] = 'A'; // Authorisation request
-        $item[] = $invoiceId; // Transaction reference - use invoice id
-        $item[] = $lastname . '_' . $firstname; // User name - not used anywhere
-        $item[] = $vaultId; // Vault Id for client card
-        $item[] = '00'; // Budget period - no budget
-        $item[] = intval( $amount * 100 ); // Transaction amount in ZA cents
-        $data[] = $item;
+    // Set data
+    $data                  = [];
+    $data['pgid']          = $payHostId;
+    $data['encryptionKey'] = $payHostSecretKey;
+    $data['reference']     = $invoiceId;
+    $data['amount']        = intval( $amount * 100 );
+    $data['currency']      = $currencyCode;
+    $data['transDate']     = $transactionDate;
+    $data['locale']        = 'en-us';
+    $data['firstName']     = $firstname;
+    $data['lastName']      = $lastname;
+    $data['email']         = $email;
+    $data['customerTitle'] = isset( $data['customerTitle'] ) ? $data['customerTitle'] : 'Mr';
+    $data['country']       = 'ZAF';
+    $data['retUrl']        = $returnUrl;
+    $data['notifyURL']     = $notifyUrl;
+    $data['recurring']     = $usePayBatch;
+    $data['vaulting']      = $vaulting;
+    $data['vaultId']       = $vaultId;
+    $payhostSoap->setData( $data );
 
-        // Use SoapClient to make requests
-        //First authorisation request
-        $soap       = $payBatchSoap->getAuthRequest( $data );
-        $wsdl       = PAYBATCHAPIWSDL;
-        $options    = ['trace' => 1, 'login' => $payBatchId, 'password' => $payBatchSecretKey];
-        $soapClient = new SoapClient( $wsdl, $options );
-        try {
-            $result = $soapClient->__soapCall( 'Auth', [
-                new SoapVar( $soap, XSD_ANYXML ),
-            ] );
-            if ( $result->Invalid == 0 ) {
-                $uploadId = $result->UploadID;
-                //Now make confirmation request to trigger actual payment attempt
-                $confirmXml    = $payBatchSoap->getConfirmRequest( $uploadId );
-                $confirmResult = $soapClient->__soapCall( 'Confirm', [
-                    new SoapVar( $confirmXml, XSD_ANYXML ),
-                ] );
-                logTransaction( $gatewayModuleName, $confirmResult, 'success' );
-            }
-        } catch ( SoapFault $fault ) {
+    $xml = $payhostSoap->getSOAP();
 
-        }
+    // Use PHP SoapClient to handle request
+    ini_set( 'soap.wsdl_cache', 0 );
+    $soapClient = new SoapClient( PAYHOSTAPIWSDL, ['trace' => 1] );
 
-    } else {
-        // A web payment request - use PayHost WebPayment request for redirect
+    try {
+        $result = $soapClient->__soapCall( 'SinglePayment', [
+            new SoapVar( $xml, XSD_ANYXML ),
+        ] );
 
-        $payhostSoap = new payhostsoap();
+        if ( array_key_exists( 'Redirect', $result->WebPaymentResponse ) ) {
+            // Redirect to Payment Portal
+            // Store key values for return response
+            $_SESSION['PAY_REQUEST_ID'] = $result->WebPaymentResponse->Redirect->UrlParams[1]->value;
+            $_SESSION['REFERENCE']      = $result->WebPaymentResponse->Redirect->UrlParams[2]->value;
 
-        // Set data
-        $data                  = [];
-        $data['pgid']          = $payHostId;
-        $data['encryptionKey'] = $payHostSecretKey;
-        $data['reference']     = $invoiceId;
-        $data['amount']        = intval( $amount * 100 );
-        $data['currency']      = $currencyCode;
-        $data['transDate']     = $transactionDate;
-        $data['locale']        = 'en-us';
-        $data['firstName']     = $firstname;
-        $data['lastName']      = $lastname;
-        $data['email']         = $email;
-        $data['customerTitle'] = isset( $data['customerTitle'] ) ? $data['customerTitle'] : 'Mr';
-        $data['country']       = 'ZAF';
-        $data['retUrl']        = $returnUrl;
-        $data['notifyURL']     = $notifyUrl;
-        $data['recurring']     = $usePayBatch;
-        $payhostSoap->setData( $data );
+            // Do redirect
+            // First check that the checksum is valid
+            $d = $result->WebPaymentResponse->Redirect->UrlParams;
 
-        $xml = $payhostSoap->getSOAP();
+            $checkSource = $d[0]->value;
+            $checkSource .= $d[1]->value;
+            $checkSource .= $d[2]->value;
+            $checkSource .= $payHostSecretKey;
+            $check = md5( $checkSource );
+            if ( $check == $d[3]->value ) {
+                $inputs = $d;
 
-        // Use PHP SoapClient to handle request
-        ini_set( 'soap.wsdl_cache', 0 );
-        $soapClient = new SoapClient( PAYHOSTAPIWSDL, ['trace' => 1] );
-
-        try {
-            $result = $soapClient->__soapCall( 'SinglePayment', [
-                new SoapVar( $xml, XSD_ANYXML ),
-            ] );
-
-            if ( array_key_exists( 'Redirect', $result->WebPaymentResponse ) ) {
-                // Redirect to Payment Portal
-                // Store key values for return response
-                $_SESSION['PAY_REQUEST_ID'] = $result->WebPaymentResponse->Redirect->UrlParams[1]->value;
-                $_SESSION['REFERENCE']      = $result->WebPaymentResponse->Redirect->UrlParams[2]->value;
-
-                // Do redirect
-                // First check that the checksum is valid
-                $d = $result->WebPaymentResponse->Redirect->UrlParams;
-
-                $checkSource = $d[0]->value;
-                $checkSource .= $d[1]->value;
-                $checkSource .= $d[2]->value;
-                $checkSource .= $payHostSecretKey;
-                $check = md5( $checkSource );
-                if ( $check == $d[3]->value ) {
-                    $inputs = $d;
-
-                    $html = <<<HTML
+                $html = <<<HTML
         <form action="{$result->WebPaymentResponse->Redirect->RedirectUrl}" method="post" name="payhost">
         <input type="hidden" name="{$inputs[0]->key}" value="{$inputs[0]->value}" />
         <input type="hidden" name="{$inputs[1]->key}" value="{$inputs[1]->value}" />
@@ -299,15 +277,12 @@ function payhostpaybatch_link( $params )
         document.forms['payhost'].submit();
 </script>
 HTML;
-
-//                    return $html;
-                }
-            } else {
-                // Process response - doesn't happen
             }
-        } catch ( SoapFault $f ) {
-            var_dump( $f );
+        } else {
+            // Process response - doesn't happen
         }
+    } catch ( SoapFault $f ) {
+        var_dump( $f );
     }
     return $html;
 }
@@ -354,7 +329,7 @@ function payhostpaybatch_refund( $params )
     $moduleName        = $params['paymentmethod'];
     $whmcsVersion      = $params['whmcsVersion'];
 
-    // perform API call to initiate refund and interpret result
+    // Perform API call to initiate refund and interpret result
 
     return array(
         // 'success' if successful, otherwise 'declined', 'error' for failure
@@ -406,593 +381,4 @@ function payhostpaybatch_cancelSubscription( $params )
         // Data to be recorded in the gateway log - can be a string or array
         'rawdata' => $responseData,
     );
-}
-
-/**
- * Class payhostsoap
- * Helper class to make SOAP call to PayHost endpoint
- */
-class payhostsoap
-{
-    /**
-     * @var string the url of the PayGate PayHost process page
-     */
-    public static $process_url = 'https://secure.paygate.co.za/PayHost/process.trans';
-
-    /**
-     * @var string the url of the PayGate PayHost WSDL
-     */
-    public static $wsdl = 'https://secure.paygate.co.za/payhost/process.trans?wsdl';
-
-    /**
-     * @var string default namespace. We add the namespace manually because of PHP's "quirks"
-     */
-    private static $ns = 'ns1';
-
-    // Standard Inputs
-    protected $pgid;
-    protected $reference;
-    protected $amount;
-    protected $currency;
-    protected $transDate;
-    protected $locale;
-    protected $payMethod;
-    protected $payMethodDetail;
-    protected $encryptionKey;
-
-    // Customer Details
-    protected $customerTitle;
-    protected $firstName;
-    protected $middleName;
-    protected $lastName;
-    protected $telephone;
-    protected $mobile;
-    protected $fax;
-    protected $email;
-    protected $dateOfBirth;
-    protected $socialSecurity;
-
-    // Address Details
-    protected $addressLine1;
-    protected $addressLine2;
-    protected $addressLine3;
-    protected $zip;
-    protected $city;
-    protected $state;
-    protected $country;
-
-    // Address checkboxes
-    protected $incCustomer = true;
-    protected $incBilling  = true;
-    protected $incShipping;
-
-    // Shipping Details
-    protected $deliveryDate;
-    protected $deliveryMethod;
-    protected $installRequired;
-
-    // Redirect Details
-    protected $retUrl;
-    protected $notifyURL;
-    protected $target;
-
-    // Risk
-    protected $riskAccNum;
-    protected $riskIpAddr;
-
-    // Airline
-    protected $ticketNumber;
-    protected $PNR;
-    protected $travellerType;
-    protected $departureAirport;
-    protected $departureCountry;
-    protected $departureCity;
-    protected $departureDateTime;
-    protected $arrivalAirport;
-    protected $arrivalCountry;
-    protected $arrivalCity;
-    protected $arrivalDateTime;
-    protected $marketingCarrierCode;
-    protected $marketingCarrierName;
-    protected $issuingCarrierCode;
-    protected $issuingCarrierName;
-    protected $flightNumber;
-
-    // Recurring orders
-    protected $recurring;
-
-    public static $DEFAULT_PGID           = '10011072130';
-    public static $DEFAULT_AMOUNT         = 3299;
-    public static $DEFAULT_CURRENCY       = 'ZAR';
-    public static $DEFAULT_LOCALE         = 'en-us';
-    public static $DEFAULT_ENCRYPTION_KEY = 'test';
-    public static $DEFAULT_TITLE          = 'Mr';
-    public static $DEFAULT_FIRST_NAME     = 'PayGate';
-    public static $DEFAULT_LAST_NAME      = 'Test';
-    public static $DEFAULT_EMAIL          = 'itsupport@paygate.co.za';
-    public static $DEFAULT_COUNTRY        = 'ZAF';
-    public static $DEFAULT_NOTIFY_URL     = 'http://www.gatewaymanagementservices.com/ws/gotNotify.php';
-    public static $DEFAULT_PAY_METHOD     = 'CC';
-
-    public function __construct()
-    {
-
-    }
-
-    public function setData( $data )
-    {
-        foreach ( $data as $key => $value ) {
-            $k        = $key;
-            $this->$k = $value;
-        }
-    }
-
-    public function getSOAP()
-    {
-        $xml = <<<XML
-<{$this::$ns}:SinglePaymentRequest>
-<{$this::$ns}:WebPaymentRequest>
-{$this->getAccount()}
-{$this->getCustomer()}
-{$this->getVault( $this->recurring )}
-{$this->getPaymentType()}
-{$this->getRedirect()}
-{$this->getOrder()}
-{$this->getRisk()}
-{$this->getUserFields()}
-</{$this::$ns}:WebPaymentRequest>
-</{$this::$ns}:SinglePaymentRequest>
-XML;
-
-        $xml = preg_replace( "/(^[\r\n]*|[\r\n]+)[\s\t]*[\r\n]+/", "\n", $xml ); // Remove empty lines to make the plain text request prettier
-
-        return $xml;
-    }
-
-    private function getVault( $recurring )
-    {
-        if ( $recurring !== true ) {
-            return '';
-        }
-
-        $token = null;
-        if ( $token == null || $token == '' ) {
-            // If token is not already stored under member then add element to request a vault transaction
-            $vault = <<<VAULT
-<!-- Vault Detail -->
-    <{$this::$ns}:Vault>true</{$this::$ns}:Vault>
-VAULT;
-            return $vault;
-        } else {
-            // Return the Vault element with valid token
-            $vault = <<<VAULT
-    <{$this::$ns}:VaultId>{$token}</{$this::$ns}:VaultId>
-VAULT;
-
-            return $vault;
-        }
-    }
-
-    private function getAccount()
-    {
-        $account = <<<XML
-<!-- Account Details -->
-    <{$this::$ns}:Account>
-    <{$this::$ns}:PayGateId>{$this->pgid}</{$this::$ns}:PayGateId>
-    <{$this::$ns}:Password>{$this->encryptionKey}</{$this::$ns}:Password>
-    </{$this::$ns}:Account>
-XML;
-
-        return $account;
-    }
-
-    private function getCustomer()
-    {
-
-        $middleName     = ( $this->middleName != '' ? "<{$this::$ns}:MiddleName>{$this->middleName}</{$this::$ns}:MiddleName>" : '' );
-        $telephone      = ( $this->telephone != '' ? "<{$this::$ns}:Telephone>{$this->telephone}</{$this::$ns}:Telephone>" : '' );
-        $mobile         = ( $this->mobile != '' ? "<{$this::$ns}:Mobile>{$this->mobile}</{$this::$ns}:Mobile>" : '' );
-        $fax            = ( $this->fax != '' ? "<{$this::$ns}:Fax>{$this->fax}</{$this::$ns}:Fax>" : '' );
-        $dateOfBirth    = ( $this->dateOfBirth != '' ? "<{$this::$ns}:DateOfBirth>{$this->dateOfBirth}</{$this::$ns}:DateOfBirth>" : '' );
-        $socialSecurity = ( $this->socialSecurity != '' ? "<{$this::$ns}:SocialSecurityNumber>{$this->socialSecurity}</{$this::$ns}:SocialSecurityNumber>" : '' );
-        $address        = ( isset( $this->incCustomer ) ? $this->getAddress() : '' );
-
-        $customer = <<<XML
-<!-- Customer Details -->
-    <{$this::$ns}:Customer>
-    <{$this::$ns}:Title>{$this->customerTitle}</{$this::$ns}:Title>
-    <{$this::$ns}:FirstName>{$this->firstName}</{$this::$ns}:FirstName>
-    {$middleName}
-    <{$this::$ns}:LastName>{$this->lastName}</{$this::$ns}:LastName>
-    {$telephone}
-    {$mobile}
-    {$fax}
-    <{$this::$ns}:Email>{$this->email}</{$this::$ns}:Email>
-    {$dateOfBirth}
-    {$socialSecurity}
-    {$address}
-    </{$this::$ns}:Customer>
-XML;
-
-        return $customer;
-    }
-
-    private function getAddress()
-    {
-
-        $address1 = ( $this->addressLine1 != '' ? "<{$this::$ns}:AddressLine>{$this->addressLine1}</{$this::$ns}:AddressLine>" : '' );
-        $address2 = ( $this->addressLine2 != '' ? "<{$this::$ns}:AddressLine>{$this->addressLine2}</{$this::$ns}:AddressLine>" : '' );
-        $address3 = ( $this->addressLine3 != '' ? "<{$this::$ns}:AddressLine>{$this->addressLine3}</{$this::$ns}:AddressLine>" : '' );
-        $city     = ( $this->city != '' ? "<{$this::$ns}:City>{$this->city}</{$this::$ns}:City>" : '' );
-        $country  = ( $this->country != '' ? "<{$this::$ns}:Country>{$this->country}</{$this::$ns}:Country>" : 'ZAF' );
-        $state    = ( $this->state != '' ? "<{$this::$ns}:State>{$this->state}</{$this::$ns}:State>" : '' );
-        $zip      = ( $this->zip != '' ? "<{$this::$ns}:Zip>{$this->zip}</{$this::$ns}:Zip>" : '' );
-
-        $address = <<<XML
-<!-- Address Details -->
-    <{$this::$ns}:Address>
-    {$address1}
-    {$address2}
-    {$address3}
-    {$city}
-    {$country}
-    {$state}
-    {$zip}
-    </{$this::$ns}:Address>
-XML;
-
-        return $address;
-    }
-
-    private function getPaymentType()
-    {
-        $paymentType = '';
-
-        if ( $this->payMethod != '' || $this->payMethodDetail != '' ) {
-            $payMethod       = ( $this->payMethod != '' ? "<{$this::$ns}:Method>{$this->payMethod}</{$this::$ns}:Method>" : '' );
-            $payMethodDetail = ( $this->payMethodDetail != '' ? "<{$this::$ns}:Detail>{$this->payMethodDetail}</{$this::$ns}:Detail>" : '' );
-
-            $paymentType = <<<XML
-<!-- Payment Type Details -->
-    <{$this::$ns}:PaymentType>
-    {$payMethod}
-    {$payMethodDetail}
-    </{$this::$ns}:PaymentType>
-XML;
-        }
-
-        return $paymentType;
-    }
-
-    private function getRedirect()
-    {
-        $target = ( isset( $this->target ) && $this->target != '' ? '<' . $this::$ns . ':Target>' . $this->target . '</' . $this::$ns . ':Target>' : '' );
-
-        $redirect = <<<XML
-<!-- Redirect Details -->
-    <{$this::$ns}:Redirect>
-    <{$this::$ns}:NotifyUrl>{$this->notifyURL}</{$this::$ns}:NotifyUrl>
-    <{$this::$ns}:ReturnUrl>{$this->retUrl}</{$this::$ns}:ReturnUrl>
-    {$target}
-    </{$this::$ns}:Redirect>
-XML;
-
-        return $redirect;
-    }
-
-    private function getBillingDetails()
-    {
-        $billing = '';
-
-        if ( isset( $this->incBilling ) ) {
-            $billing = <<<XML
-<{$this::$ns}:BillingDetails>
-{$this->getCustomer()}
-{$this->getAddress()}
-</{$this::$ns}:BillingDetails>
-XML;
-        }
-
-        return $billing;
-    }
-
-    private function getShippingDetails()
-    {
-        $shipping = '';
-
-        if ( isset( $this->incShipping ) || $this->deliveryDate != '' || $this->deliveryMethod != '' || isset( $this->installRequired ) ) {
-
-            $address         = ( isset( $this->incShipping ) ? $this->getAddress() : '' );
-            $deliveryDate    = ( $this->deliveryDate != '' ? "<{$this::$ns}:DeliveryDate>{$this->deliveryDate}</{$this::$ns}:DeliveryDate>" : '' );
-            $deliveryMethod  = ( $this->deliveryMethod != '' ? "<{$this::$ns}:DeliveryMethod>{$this->deliveryMethod}</{$this::$ns}:DeliveryMethod>" : '' );
-            $installRequired = ( $this->installRequired != '' ? "<{$this::$ns}:InstallationRequested>{$this->installRequired}</{$this::$ns}:InstallationRequested>" : '' );
-
-            $shipping = <<<XML
-<{$this::$ns}:ShippingDetails>
-{$this->getCustomer()}
-{$address}
-{$deliveryDate}
-{$deliveryMethod}
-{$installRequired}
-</{$this::$ns}:ShippingDetails>
-XML;
-        }
-
-        return $shipping;
-    }
-
-    private function getOrder()
-    {
-
-        $order = <<<XML
-<!-- Order Details -->
-    <{$this::$ns}:Order>
-    <{$this::$ns}:MerchantOrderId>{$this->reference}</{$this::$ns}:MerchantOrderId>
-    <{$this::$ns}:Currency>{$this->currency}</{$this::$ns}:Currency>
-    <{$this::$ns}:Amount>{$this->amount}</{$this::$ns}:Amount>
-    <{$this::$ns}:TransactionDate>{$this->transDate}</{$this::$ns}:TransactionDate>
-    {$this->getBillingDetails()}
-    {$this->getShippingDetails()}
-    {$this->getAirlineFields()}
-    <{$this::$ns}:Locale>{$this->locale}</{$this::$ns}:Locale>
-    </{$this::$ns}:Order>
-XML;
-
-        return $order;
-    }
-
-    private function getRisk()
-    {
-        $risk = '';
-
-        if ( $this->riskAccNum != '' && $this->riskIpAddr != '' ) {
-            $risk = <<<XML
-<!-- Risk Details -->
-<{$this::$ns}:Risk>
-<{$this::$ns}:AccountNumber>{$this->riskAccNum}</{$this::$ns}:AccountNumber>
-<{$this::$ns}:IpV4Address>{$this->riskIpAddr}</{$this::$ns}:IpV4Address>
-</{$this::$ns}:Risk>
-XML;
-        }
-
-        return $risk;
-    }
-
-    private function getUserFields()
-    {
-
-        $userDefined = '<!-- User Fields -->' . PHP_EOL;
-        $i           = 1;
-
-        while ( $i >= 1 ) {
-            if ( isset( $this->{'userKey' . $i} ) && $this->{'userKey' . $i} != '' && isset( $this->{'userField' . $i} ) && $this->{'userField' . $i} != '' ) {
-
-                $key   = $this->{'userKey' . $i};
-                $value = $this->{'userField' . $i};
-
-                $userDefined
-                .= <<<XML
-    <{$this::$ns}:UserDefinedFields>
-    <{$this::$ns}:key>{$key}</ns1:key>
-    <{$this::$ns}:value>{$value}</ns1:value>
-    </{$this::$ns}:UserDefinedFields>
-
-XML;
-                $i++;
-            } else {
-                break;
-            }
-        }
-
-        return $userDefined;
-    }
-
-    private function getPassenger()
-    {
-        $middleName     = ( $this->middleName != '' ? "<{$this::$ns}:MiddleName>{$this->middleName}</{$this::$ns}:MiddleName>" : '' );
-        $telephone      = ( $this->telephone != '' ? "<{$this::$ns}:Telephone>{$this->telephone}</{$this::$ns}:Telephone>" : '' );
-        $mobile         = ( $this->mobile != '' ? "<{$this::$ns}:Mobile>{$this->mobile}</{$this::$ns}:Mobile>" : '' );
-        $fax            = ( $this->fax != '' ? "<{$this::$ns}:Fax>{$this->fax}</{$this::$ns}:Fax>" : '' );
-        $dateOfBirth    = ( $this->dateOfBirth != '' ? "<{$this::$ns}:DateOfBirth>{$this->dateOfBirth}</{$this::$ns}:DateOfBirth>" : '' );
-        $socialSecurity = ( $this->socialSecurity != '' ? "<{$this::$ns}:SocialSecurityNumber>{$this->socialSecurity}</{$this::$ns}:SocialSecurityNumber>" : '' );
-
-        $passenger = <<<XML
-<{$this::$ns}:Passenger>
-<{$this::$ns}:Title>{$this->customerTitle}</{$this::$ns}:Title>
-<{$this::$ns}:FirstName>{$this->firstName}</{$this::$ns}:FirstName>
-{$middleName}
-<{$this::$ns}:LastName>{$this->lastName}</{$this::$ns}:LastName>
-{$telephone}
-{$mobile}
-{$fax}
-<{$this::$ns}:Email>{$this->email}</{$this::$ns}:Email>
-{$dateOfBirth}
-{$socialSecurity}
-</{$this::$ns}:Passenger>
-
-XML;
-
-        return $passenger;
-    }
-
-    private function getFlightLegs()
-    {
-        $flightLeg = <<<XML
-<{$this::$ns}:FlightLegs>
-<{$this::$ns}:DepartureAirport>{$this->departureAirport}</{$this::$ns}:DepartureAirport>
-<{$this::$ns}:DepartureCountry>{$this->departureCountry}</{$this::$ns}:DepartureCountry>
-<{$this::$ns}:DepartureCity>{$this->departureCity}</{$this::$ns}:DepartureCity>
-<{$this::$ns}:DepartureDateTime>{$this->departureDateTime}</{$this::$ns}:DepartureDateTime>
-<{$this::$ns}:ArrivalAirport>{$this->arrivalAirport}</{$this::$ns}:ArrivalAirport>
-<{$this::$ns}:ArrivalCountry>{$this->arrivalCountry}</{$this::$ns}:ArrivalCountry>
-<{$this::$ns}:ArrivalCity>{$this->arrivalCity}</{$this::$ns}:ArrivalCity>
-<{$this::$ns}:ArrivalDateTime>{$this->arrivalDateTime}</{$this::$ns}:ArrivalDateTime>
-<{$this::$ns}:MarketingCarrierCode>{$this->marketingCarrierCode}</{$this::$ns}:MarketingCarrierCode>
-<{$this::$ns}:MarketingCarrierName>{$this->marketingCarrierName}</{$this::$ns}:MarketingCarrierName>
-<{$this::$ns}:IssuingCarrierCode>{$this->issuingCarrierCode}</{$this::$ns}:IssuingCarrierCode>
-<{$this::$ns}:IssuingCarrierName>{$this->issuingCarrierName}</{$this::$ns}:IssuingCarrierName>
-<{$this::$ns}:FlightNumber>{$this->flightNumber}</{$this::$ns}:FlightNumber>
-<{$this::$ns}:BaseFareAmount>{$this->amount}</{$this::$ns}:BaseFareAmount>
-<{$this::$ns}:BaseFareCurrency>{$this->currency}</{$this::$ns}:BaseFareCurrency>
-</{$this::$ns}:FlightLegs>
-XML;
-
-        return $flightLeg;
-    }
-
-    private function getAirlineFields()
-    {
-        $airline = '';
-
-        if ( $this->PNR != '' ) {
-            $airline = <<<XML
-<{$this::$ns}:AirlineBookingDetails>
-<{$this::$ns}:TicketNumber>{$this->ticketNumber}</{$this::$ns}:TicketNumber>
-<{$this::$ns}:PNR>{$this->PNR}</{$this::$ns}:PNR>
-<{$this::$ns}:Passengers>
-{$this->getPassenger()}
-<{$this::$ns}:TravellerType>{$this->travellerType}</{$this::$ns}:TravellerType>
-</{$this::$ns}:Passengers>
-{$this->getFlightLegs()}
-</{$this::$ns}:AirlineBookingDetails>
-XML;
-        }
-
-        return $airline;
-    }
-
-    public function getSOAPData()
-    {
-        $data                                 = [];
-        $data['WebPaymentRequest']            = [];
-        $data['WebPaymentRequest']['Account'] = $this->getAccountData();
-
-        return $data;
-    }
-
-    private function getAccountData()
-    {
-        $PayGateId = $this->pgid;
-        $Password  = $this->encryptionKey;
-        return ['PayGateId' => $PayGateId, 'Password' => $Password];
-    }
-}
-
-/**
- * Class paybatchsoap
- * Helper class for making SOAP requests to PayBatch API
- */
-class paybatchsoap
-{
-    /**
-     * @var string the url of the PayGate PayBatch process page
-     */
-    public static $process_url = 'https://secure.paygate.co.za/paybatch/1.2/process.trans';
-
-    /**
-     * @var string the url of the PayGate PayBatch WSDL
-     */
-    public static $wsdl = 'https://secure.paygate.co.za/paybatch/1.2/PayBatch.wsdl';
-
-    /**
-     * @var string default namespace. We add the namespace manually because of PHP's "quirks"
-     */
-    private static $ns = 'ns1';
-
-    /**
-     * @var string $notifyUrl
-     */
-    private static $notifyUrl;
-
-    /**
-     * @var string $soapStart , $soapEnd
-     * SOAP HEADERS
-     */
-    private static $soapStart = '<SOAP-ENV:Envelope xmlns:SOAP-ENV="http://schemas.xmlsoap.org/soap/envelope/">
-    <SOAP-ENV:Header/>
-    <SOAP-ENV:Body>';
-    private static $soapend = '</SOAP-ENV:Body>
-</SOAP-ENV:Envelope>';
-
-    /**
-     * @var array of data for batchline
-     */
-    protected $batchData = [];
-
-    protected $batchReference = 'PayBatch_';
-
-    public function __construct( $notifyUrl )
-    {
-        $this->batchReference .= date( 'Y-m-d' ) . '_' . uniqid();
-        $this::$notifyUrl = $notifyUrl;
-    }
-
-    /**
-     * @param $data array of batchline type
-     */
-    public function setBatchData( $data )
-    {
-        foreach ( $data as $line ) {
-            $this->batchData[] = $line;
-        }
-    }
-
-    /**
-     * @param $data input data array
-     */
-    public function getAuthRequest( $data )
-    {
-        $this->setBatchData( $data );
-        try {
-            // Use SimpleXMLElement to build structure
-            $xml = new SimpleXMLElement( '<Auth />' );
-            $xml->addChild( 'BatchReference', $this->batchReference );
-            $xml->addChild( 'NotificationUrl', $this::$notifyUrl );
-
-            $batchData = $xml->addChild( 'BatchData' );
-            foreach ( $this->batchData as $line ) {
-                $batchLine = '';
-                foreach ( $line as $item ) {
-                    $batchLine .= $item . ',';
-                }
-                $batchLine = rtrim( $batchLine, ',' );
-                $batchData->addChild( 'BatchLine', $batchLine );
-            }
-
-            // Remove XML headers
-            $dom = new DOMDocument();
-            $dom->loadXML( $xml->asXML() );
-
-            $soap = $dom->saveXML( $dom->documentElement );
-
-            // Remove Auth tag - added in __soapCall
-            $childrenOnly = str_replace( ['<Auth>', '</Auth>'], '', $soap );
-
-            return $childrenOnly;
-        } catch ( Exception $e ) {
-            return $e->getMessage();
-        }
-    }
-
-    public function getConfirmRequest( $uploadId )
-    {
-        try {
-            //use SimpleXmlElement for better control of children
-            $xml = new SimpleXMLElement( '<Query />' );
-
-            $xml->addChild( 'UploadID', $uploadId );
-
-            //use DomDocument to remove XML headers
-            $dom = new DOMDocument();
-            $dom->loadXML( $xml->asXML() );
-
-            $soap = $dom->saveXML( $dom->documentElement );
-
-            //Remove Confirm tag because we pass it in the __soapCall
-            $childrenOnly = str_replace( ['<Query>', '</Query>'], '', $soap );
-            return $childrenOnly;
-        } catch ( Exception $e ) {
-            return $e->getMessage();
-        }
-    }
 }
